Move the edited file to the correct location
